+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Rename `EveOnline::ESI::UniverseSystems#universe_systems_ids` to `#universe_system_ids`
 * Rename `EveOnline::ESI::UniverseRegions#universe_regions_ids` to `#universe_region_ids`
 * Rename `EveOnline::ESI::UniverseConstellations#constellations_ids` to `#constellation_ids`
+* Rename `EveOnline::ESI::DogmaAttributes#attributes` to `#attribute_ids`
 
 ## v0.18.0